Release 1.0.15

+++ src/client/chat/shared/contact-action/contact-actions.component.ts
@@ -1,37 +1,51 @@
-import { Component, OnInit, Input, ViewChild } from '@angular/core';
+import { Component, OnInit, Input, ViewChild, EventEmitter, Output } from '@angular/core';
+import { Router } from '@angular/router';
+
 import { ChatService } from '../services/chat.service';
 import { ZChatShareAddToConversationComponent } from '../modal/add-to-conversation.component';
+import { ConfirmationService } from 'primeng/components/common/api';
+import { ZoneReportService } from '../../../core/shared/form/report/report.service';
 
-declare let $:any;
+declare let $: any;
 
 @Component({
   moduleId: module.id,
   selector: 'z-chat-contact-actions',
-  templateUrl: 'contact-actions.component.html'
+  templateUrl: 'contact-actions.component.html',
+  styles: [':host{display: inline-block;}']
 })
 export class ZChatContactActionsComponent implements OnInit {
-  @Input() contact:any;
-  conversationUrl:any;
+  @Input() contact: any;
+  conversationUrl: any;
+  profileUrl: any;
   @ViewChild('addConversation') addConversation: ZChatShareAddToConversationComponent;
+  @Output() updateEvent: EventEmitter<any> = new EventEmitter<any>();
   // Config component
-  @Input() config:any = {
+  @Input() config: any = {
     history: false
   };
 
-  constructor(private chatService: ChatService) {
+  constructor(private router: Router,
+              private chatService: ChatService,
+              private zoneReportService: ZoneReportService,
+              private confirmationService: ConfirmationService) {
     this.conversationUrl = this.chatService.constant.conversationUrl;
+    this.profileUrl = this.chatService.constant.profileUrl;
   }
 
   ngOnInit() {
-  //   console.log(this.config)
+    //   console.log(this.config)
   }
 
-  onSelect(contact:any) {
-    if(this.config.history) {
+  onSelect(contact: any) {
+    console.log(contact);
+    if (this.config.history) {
       this.chatService.updateHistory(contact);
     }
     $('#chat-message-text').focus();
     this.chatService.selectContact(contact);
+
+    this.router.navigate([this.conversationUrl, contact.id]);
   }
 
   addToConversation() {
@@ -39,19 +53,41 @@ export class ZChatContactActionsComponent implements OnInit {
   }
 
   addBlackList() {
-    // console.log(this.contact);
-    this.chatService.addGroupUserBlackList(this.contact);
+    this.confirmationService.confirm({
+      message: 'Are you sure you want to add this contact to black list ?',
+      header: 'Add To Black List',
+      accept: () => {
+        this.chatService.addGroupUserBlackList(this.contact.partner_id);
+      }
+    });
   }
 
   disableNotification() {
     this.chatService.updateNotification(this.contact, {notification: false});
+    this.contact.notification = !this.contact.notification;
   }
 
   enableNotification() {
     this.chatService.updateNotification(this.contact, {notification: true});
+    this.contact.notification = !this.contact.notification;
   }
 
   deleteContact() {
-    this.chatService.deleteContact(this.contact);
+    this.confirmationService.confirm({
+      message: 'Are you sure you want to delete this contact ?',
+      header: 'Delete Contact',
+      accept: () => {
+        this.chatService.deleteContact(this.contact);
+        this.updateEvent.emit();
+      }
+    });
+  }
+
+  onFavorite() {
+    this.chatService.addGroupUserFavorite(this.contact);
+  }
+
+  report(uuid: any) {
+    this.zoneReportService.friend(uuid);
   }
 }
